Move user profile endpoints to dedicated profile module
