+++ apps/backend/src/me/profile/profile.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ProfileController } from './controller/profile.controller';
+import { UserModule } from 'src/user/user.module';
+
+@Module({
+  imports: [UserModule],
+  controllers: [ProfileController],
+})
+export class ProfileModule {}
